requirements.txt
pyyaml==4.2b4
requests==2.21.0
requests-toolbelt==0.9.1
slackbot==0.5.3
slacker==0.12.0
urllib3==1.24.1
websocket-client==0.44.0
